Task 14

+++ 14-binary_tree_balance.c
@@ -50,5 +50,5 @@ int binary_tree_balance(const binary_tree_t *tree)
 
 	left_h = binary_tree_height(tree->left);
 
-	return ((int)(right_h - left_h);
+	return (right_h - left_h);
 }
